# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Yeungjin-Global",
        "description": "YJG API Documentation",
        "version": "0.1"
    },
    "paths": {
...
        "/api/restaurant/semester": {
            "post": {
                "tags": [
                    "식수"
                ],
                "summary": "식수 학기 신청",
                "description": "식수 학기 신청을 처리합니다",
                "operationId": "4aad784a36b64dea17b0450cf35fa5b2",
                "requestBody": {
                    "description": "학생 식사 신청 정보",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "payment": {
                                        "description": "입금 확인",
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Fail"
                    }
                }
            }
        },
        "/api/restaurant/semester/g/payment{id}": {
            "get": {
                "tags": [
                    "식수"
                ],
                "summary": "식수 학기 신청 입금여부",
                "description": "식수 학기 신청의 입금여부를 확인 합니다",
                "operationId": "cad43b7b850c5562c601f66a44414c1a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "조회 할 식수 신청 아이디",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Fail"
                    }
                }
            }
        },
        "/api/restaurant/semester/p/payment/{id}": {
            "post": {
                "tags": [
                    "식수"
                ],
                "summary": "학기 입금여부 수정",
                "description": "학기 입금여부를 수정",
                "operationId": "5625fe8196b922a488b72c7af3cf9d15",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "확인할 식수신청 id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "수정할 입금여부(true,false), 신청id",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "payment": {
                                        "description": "입금여부",
                                        "type": "boolean",
                                        "example": "true"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Validation Exception"
                    },
                    "500": {
                        "description": "Fail"
                    }
                }
            }
        },
        "/api/restaurant/semester/delete/{id}": {
            "delete": {
                "tags": [
                    "식수"
                ],
                "summary": "학기 식수 신청 삭제",
                "description": "학기 식수 신청 삭제",
                "operationId": "2a0bf69d8561082237eeb76058ed04ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "삭제할 학기 식수 신청 아이디",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Fail"
                    }
                }
            }
        },
        "/api/restaurant/semester/show": {
            "get": {
                "tags": [
                    "식수"
                ],
                "summary": "학기 식수 신청 리스트 보기, 검색",
                "description": "학기 식수 신청 리스트 보기, 검색하기",
                "operationId": "5adfde227d85d6dd22f2d3461f192e8c",
                "requestBody": {
                    "description": "찾고싶은 학생 이름",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "학생이름",
                                        "type": "string",
                                        "example": "권지훈"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Fail"
                    }
                }
            }
        },
        "/api/restaurant/semester/show/user": {
            "get": {
                "tags": [
                    "식수"
                ],
                "summary": "학기 식수 유저 정보",
                "description": "학기 식수 유조 정보 확인",
                "operationId": "e28765ed7d54a2d28b8aa78b05cd511d",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Fail"
                    }
                }
            }
        },
        "/api/restaurant/semester/show/user/after": {
            "get": {
                "tags": [
                    "식수"
                ],
                "summary": "학기 식수 유저 정보(신청 후)",
                "description": "학기 식수 유조 정보 확인(신청 후)",
                "operationId": "6ccda42e9859f785e8f3100006436ecc",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Fail"
                    }
                }
            }
        },
...
    }
}